Changes to shop functionality
Seperated shop crate handling to new file.

Redone Crate Load & Save - will need to reset your profile Var
'Neb_telecache'.

All items placed in side crate are now dismantled to allow for more
reliable saving, both client and remotes.

Removed open crate from all BUY scripts as discussed to allow continued
buying - if something gets added to the crate it is still placed in
front of you as a reminder.

Changed when buying attachments they are now placed on the players
current weapon if compatible.

When buying a new weapon any old attachments are transfered if it dose
not come with any in the equivilant slots.